Migration with Local Disks Mirroring: Added a new '-q' flag to xl migrate for disk mirorring
When the '-q' flag is provided to the 'xl migrate' command all the block devices
that are visible to QEMU (by executing QMP qmp_query_blocks) should be mirrored
to the destination node. If the flag is not present migration flow will be
equivalent to current migration flow. A new 'mirror_disks' field is added to the
xl domain_create struct to indicate that disks will be mirrored during instance
startup. A new LIBXL_SUSPEND_MIRROR_DISK flag is added to indicate that the
block mirror jobs should be performed during the suspension of the instance.
libxl_domain_suspend takes a new 'hostname' param with the name of the host
where the QEMU drives will be mirrored to. libxl_domain_suspend takes a
'recv_fd' param that is used for receving messages from destination during
migration.

Signed-off-by: Bruno Alvisio <bruno.alvisio@gmail.com>