Local disk migration: Added new QMP commands used for local disk migration
Local disk migration uses the embedded NBD server in QEMU. In order to achive
this, an NBD server is added at instance creation time in the destination
(nbd_server_add).
A drive mirror disk command (driver_mirror) is executed on the source to
replicate the disk and forward all the writes from the source to the
destination.
When the local disk migration option is used the QEMU process will be started
with the "-incoming defer" option instead of "-incoming fd". Once the disk and
memory has been trasnferred to the destination the QEMU process will be started
using the migrate incoming command.
For migration it will be necessary to track the status of block devices and
block jobs (mirroring is a type of block job).

Signed-off-by: Bruno Alvisio <bruno.alvisio@gmail.com>