Local Storage Migration: Added support in libxl to handle QMP events
Local storage migration uses the QEMU NBD server. To migrate the disk, a 'disk
mirror job' is started so that the block devices emulated by QEMU are mirrored
(and sync) to the destination node. Once the mirroring job is done and both the
disk are in sync in source and destination, QEMU sends an asynchronous QMP
event. This codes add support to register handlers when QMP events are received.

Signed-off-by: Bruno Alvisio <bruno.alvisio@gmail.com>