Local Storage Migration: Added QMP commands used for local disk migration
Local disk migration uses the embedded NBD server in QEMU. A NBD server is added
in libxl during instance creation time in the destination (QMP command:
nbd_server_add). The drive mirror command (QMP: driver_mirror) is executed on the
source during 'instance save' to replicate the disk in the destination. When the
local disk migration option is used the QEMU process will be started with the
"-incoming defer" option instead of "-incoming fd". Once the disk and memory
have been transferred to the destination the QEMU process will be resumed using
the migrate incoming command (QMP command: migrate_incoming). For local disk
migration it will be necessary to query and track the status of block devices
and block jobs. (QMP: query_block, qurey_block_jobs)

Signed-off-by: Bruno Alvisio <bruno.alvisio@gmail.com>