Local Storage Migration: QEMU process is started with '-incoming defer' option
For the local disk migration scenario, the QEMU process is started before the
virtual RAM is transferred to the destination node. Instead, it is just started
so that the QEMU embedded NBD server and mirror job can be started. Once the
QEMU state is transferred, the QEMU process wil be started using the QMP
'migrate-incoming' command.

When qemu-xen is started with '-incoming defer' option the initial state of the
instance will be 'inmigrate'. Thus, when either 'running' or 'inmigrate' are
read from xenstore the spawn detach is carried out.

Signed-off-by: Bruno Alvisio <bruno.alvisio@gmail.com>